refactor: 가독성을 높이기 위해 EventTrigger와 실행되는 함수 모듈 분리
