+++ src/event/triggerClickEvent.js
@@ -1,13 +1,13 @@
 import { countTodoItem } from '../todoList/countTodoItem.js';
 import {
-  changCompletedTodoItem,
   removeTodoItem,
+  toggleCompletedTodoItem,
 } from '../todoList/setTodoItem.js';
 import { hashChangeEvent } from './hashChangeEvent.js';
 
 export const triggerClickEvent = ({ target }) => {
   if (target.className === 'toggle') {
-    changCompletedTodoItem(target.closest('li'));
+    toggleCompletedTodoItem(target.closest('li'));
   }
   if (target.className === 'destroy') {
     removeTodoItem(target.closest('li'));
