+++ src/todoList/setTodoItem.js
@@ -1,4 +1,4 @@
-export const changCompletedTodoItem = (todoItem) => {
+export const toggleCompletedTodoItem = (todoItem) => {
   const $checkbox = todoItem.querySelector('input');
 
   if (todoItem.className === '') {
@@ -11,4 +11,20 @@ export const changCompletedTodoItem = (todoItem) => {
 
 export const removeTodoItem = (todoItem) => {
   return todoItem.remove();
+};
+
+const exitEditTodoItem = ({ target, key }) => {
+  const todoItem = target.closest('li');
+  if (key === 'Escape') {
+    todoItem.removeAttribute('class');
+  }
+  if (key === 'Enter') {
+    todoItem.removeAttribute('class');
+    todoItem.querySelector('label').innerHTML = target.value;
+  }
+};
+
+export const editTodoItem = (todoItem) => {
+  todoItem.setAttribute('class', 'editing');
+  todoItem.addEventListener('keyup', exitEditTodoItem);
 };
